Inline abstract superclass
