+++ driver-core/src/main/com/mongodb/internal/connection/ByteBufBsonDocument.java
@@ -21,10 +21,13 @@ import org.bson.BsonBinaryReader;
 import org.bson.BsonDocument;
 import org.bson.BsonReader;
 import org.bson.BsonType;
+import org.bson.BsonValue;
 import org.bson.ByteBuf;
 import org.bson.RawBsonDocument;
 import org.bson.codecs.BsonDocumentCodec;
+import org.bson.codecs.BsonValueCodecProvider;
 import org.bson.codecs.DecoderContext;
+import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.io.ByteBufferBsonInput;
 import org.bson.json.JsonMode;
 import org.bson.json.JsonWriter;
@@ -34,10 +37,21 @@ import java.io.InvalidObjectException;
 import java.io.ObjectInputStream;
 import java.io.StringWriter;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+import java.util.Set;
 
-class ByteBufBsonDocument extends AbstractByteBufBsonDocument {
+import static com.mongodb.assertions.Assertions.assertNotNull;
+import static com.mongodb.assertions.Assertions.notNull;
+import static org.bson.codecs.BsonValueCodecProvider.getClassForBsonType;
+import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
+
+class ByteBufBsonDocument extends BsonDocument {
     private static final long serialVersionUID = 2L;
+
+    private static final CodecRegistry REGISTRY = fromProviders(new BsonValueCodecProvider());
 
     private final transient ByteBuf byteBuf;
 
@@ -99,6 +113,14 @@ class ByteBufBsonDocument extends AbstractByteBufBsonDocument {
         return new BsonBinaryReader(new ByteBufferBsonInput(byteBuf.duplicate()));
     }
 
+    @SuppressWarnings("MethodDoesntCallSuperMethod")
+    @Override
+    public BsonDocument clone() {
+        byte[] clonedBytes = new byte[byteBuf.remaining()];
+        byteBuf.get(byteBuf.position(), clonedBytes);
+        return new RawBsonDocument(clonedBytes);
+    }
+
     @Nullable
     <T> T findInDocument(final Finder<T> finder) {
         ByteBuf duplicateByteBuf = byteBuf.duplicate();
@@ -118,14 +140,6 @@ class ByteBufBsonDocument extends AbstractByteBufBsonDocument {
         return finder.notFound();
     }
 
-    @SuppressWarnings("MethodDoesntCallSuperMethod")
-    @Override
-    public BsonDocument clone() {
-        byte[] clonedBytes = new byte[byteBuf.remaining()];
-        byteBuf.get(byteBuf.position(), clonedBytes);
-        return new RawBsonDocument(clonedBytes);
-    }
-
     int getSizeInBytes() {
         return byteBuf.getInt(byteBuf.position());
     }
@@ -141,6 +155,197 @@ class ByteBufBsonDocument extends AbstractByteBufBsonDocument {
 
     ByteBufBsonDocument(final ByteBuf byteBuf) {
         this.byteBuf = byteBuf;
+    }
+
+    @Override
+    public void clear() {
+        throw new UnsupportedOperationException("ByteBufBsonDocument instances are immutable");
+    }
+
+    @Override
+    public BsonValue put(final String key, final BsonValue value) {
+        throw new UnsupportedOperationException("ByteBufBsonDocument instances are immutable");
+    }
+
+    @Override
+    public BsonDocument append(final String key, final BsonValue value) {
+        throw new UnsupportedOperationException("ByteBufBsonDocument instances are immutable");
+    }
+
+    @Override
+    public void putAll(final Map<? extends String, ? extends BsonValue> m) {
+        throw new UnsupportedOperationException("ByteBufBsonDocument instances are immutable");
+    }
+
+    @Override
+    public BsonValue remove(final Object key) {
+        throw new UnsupportedOperationException("ByteBufBsonDocument instances are immutable");
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return assertNotNull(findInDocument(new Finder<Boolean>() {
+            @Override
+            public Boolean find(final BsonReader bsonReader) {
+                return false;
+            }
+
+            @Override
+            public Boolean notFound() {
+                return true;
+            }
+        }));
+    }
+
+    @Override
+    public int size() {
+        return assertNotNull(findInDocument(new Finder<Integer>() {
+            private int size;
+
+            @Override
+            @Nullable
+            public Integer find(final BsonReader bsonReader) {
+                size++;
+                bsonReader.readName();
+                bsonReader.skipValue();
+                return null;
+            }
+
+            @Override
+            public Integer notFound() {
+                return size;
+            }
+        }));
+    }
+
+    @Override
+    public Set<Entry<String, BsonValue>> entrySet() {
+        return toBaseBsonDocument().entrySet();
+    }
+
+    @Override
+    public Collection<BsonValue> values() {
+        return toBaseBsonDocument().values();
+    }
+
+    @Override
+    public Set<String> keySet() {
+        return toBaseBsonDocument().keySet();
+    }
+
+    @Override
+    public boolean containsKey(final Object key) {
+        if (key == null) {
+            throw new IllegalArgumentException("key can not be null");
+        }
+
+        Boolean containsKey = findInDocument(new Finder<Boolean>() {
+            @Override
+            public Boolean find(final BsonReader bsonReader) {
+                if (bsonReader.readName().equals(key)) {
+                    return true;
+                }
+                bsonReader.skipValue();
+                return null;
+            }
+
+            @Override
+            public Boolean notFound() {
+                return false;
+            }
+        });
+        return containsKey != null ? containsKey : false;
+    }
+
+    @Override
+    public boolean containsValue(final Object value) {
+        Boolean containsValue = findInDocument(new Finder<Boolean>() {
+            @Override
+            public Boolean find(final BsonReader bsonReader) {
+                bsonReader.skipName();
+                if (deserializeBsonValue(bsonReader).equals(value)) {
+                    return true;
+                }
+                return null;
+            }
+
+            @Override
+            public Boolean notFound() {
+                return false;
+            }
+        });
+        return containsValue != null ? containsValue : false;
+    }
+
+    @Nullable
+    @Override
+    public BsonValue get(final Object key) {
+        notNull("key", key);
+        return findInDocument(new Finder<BsonValue>() {
+            @Override
+            public BsonValue find(final BsonReader bsonReader) {
+                if (bsonReader.readName().equals(key)) {
+                    return deserializeBsonValue(bsonReader);
+                }
+                bsonReader.skipValue();
+                return null;
+            }
+
+            @Nullable
+            @Override
+            public BsonValue notFound() {
+                return null;
+            }
+        });
+    }
+
+    /**
+     * Gets the first key in this document.
+     *
+     * @return the first key in this document
+     * @throws java.util.NoSuchElementException if the document is empty
+     */
+    public String getFirstKey() {
+        return assertNotNull(findInDocument(new Finder<String>() {
+            @Override
+            public String find(final BsonReader bsonReader) {
+                return bsonReader.readName();
+            }
+
+            @Override
+            public String notFound() {
+                throw new NoSuchElementException();
+            }
+        }));
+    }
+
+    interface Finder<T> {
+        @Nullable
+        T find(BsonReader bsonReader);
+        @Nullable
+        T notFound();
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        return toBaseBsonDocument().equals(o);
+    }
+
+    @Override
+    public int hashCode() {
+        return toBaseBsonDocument().hashCode();
+    }
+
+    private BsonValue deserializeBsonValue(final BsonReader bsonReader) {
+        return REGISTRY.get(getClassForBsonType(bsonReader.getCurrentBsonType())).decode(bsonReader, DecoderContext.builder().build());
+    }
+
+    // see https://docs.oracle.com/javase/6/docs/platform/serialization/spec/output.html
+    Object writeReplace() {
+        return toBaseBsonDocument();
     }
 
     // see https://docs.oracle.com/javase/6/docs/platform/serialization/spec/input.html
